separate wsserver and events
