+++ server/wsserver.ts
@@ -3,68 +3,8 @@ import { Message, User } from "./models"
 import { IMessage, IUser } from "../common/data";
 import { server } from "./server";
 import { ReceiveEventType, SendEventType } from "../common/eventType" 
+import { newMessage, BaseReceiveEvent, JoinEvent, CreateMessageEvent, DeleteMessageEvent, LeftEvent } from "./events"
 
-let connection_number = 0
-
-abstract class BaseReceiveEvent {
-  constructor(protected user: IUser, protected ev?: ReceiveEventType, protected value?: string) {
-  }
-  abstract response(target: Function): void
-}
-class CreateMessageEvent extends BaseReceiveEvent{
-  response(target: Function) {
-    if (this.value === "")
-      return
-        
-    User.findOne({
-      where: {
-        name: this.user.name
-      },
-      raw: true
-    }).then((user) => {
-      Message.create({
-        text: this.value,
-        // omg wtf...
-        userId: (<any>user)["id"]
-      })
-    })
-
-    target(newMessage(SendEventType.NEW_MESSAGE, {
-      text: this.value,
-      user: {
-        name: this.user.name
-      }
-    }))
-  }
-}
-class DeleteMessageEvent extends BaseReceiveEvent{
-  response(target: Function) {
-    Message.destroy({where: {id: this.value}})
-    return target(newMessage(SendEventType.DELETE_MESSAGE, this.value) )
-  }
-}
-class JoinEvent extends BaseReceiveEvent {
-  response(target: Function) {
-    connection_number += 1
-    return target(newMessage(SendEventType.USER_JOIN, {
-      "connections": connection_number
-    }))
-  }
-}
-class LeftEvent extends BaseReceiveEvent {
-  response(target: Function) {
-    connection_number -= 1
-    return target(newMessage(SendEventType.USER_LEAVE, {
-      "connections": connection_number
-    }))
-  }
-}
-let newMessage = (ev: SendEventType, value: any) => {
-  return JSON.stringify({
-    ev: SendEventType[ev],
-    value: value,
-  })
-}
 let wss = new Server({server})
 
 let broadcast = (message: string): void => {
